WordCope/requirements.txt
pymongo
pyyaml
requests
demjson
jieba
Cython
beautifulsoup4
fasttext
numpy
flask
scipy
sklearn